requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

beautifulsoup4==4.12.3
# need boto3 ver 1.35.99 because of bug in Yandex Object Storage
boto3==1.35.99
botocore==1.35.99
colored==2.2.5
python-dotenv==1.0.1
python3-discogs-client==2.7.1
requests==2.32.3
yandex-cloud-ml-sdk==0.3.0
yandex-music==2.2.0
sqlalchemy==2.0.41
psycopg2==2.9.10
pillow==11.3.0